Snap to the closest edge when velocity is 0
When user slowly moves the PiP from one edge to the other, there could
be chances that VelocityTrack reports 0 in x-velocity and used to snap
to the right edge, which may not be intuitive to the user if the PiP
window has passed the midpoint.

In this change, we compare the centerX of the PiP window against the
display centerX, and added a boost in velocity towards nearest edge.

Updated also the pip2 counterpart.

Flag: EXEMPT bugfix
Video: http://recall/-/aaaaaabFQoRHlzixHdtY/bnpoNB1sS8Dujtfb5xUKkY
Bug: 323102290
Test: Slowly moves PiP from left to right side, and vice versa
Change-Id: I3b7c4ac14415d75164ba81f8ef471f5cae3f5ed6