camera2: allow access to CaptureRequest.mLogicalCameraSettings
Commit 4df0f1ef710d1f41d0a5f8dd7e3e4efc20bd985f
("Add maxTargetSdk restriction to unused APIs.") added a
maxTargetSdk to the @UnsupportedAppUsage annotation, disallowing
apps with a target SDK greater than 30 to make use of APIs
that were already unsupported for app usage (except system apps
with a whitelist).

The commit was made under the presumption that there's no
evidence that these APIs are used.

One of the hidden APIs, CaptureRequest.mLogicalCameraSettings,
is used by Motorola's camera app, MotCamera4, which has a
targetSdkVersion of 33/34.

Remove the target SDK restriction for the member to allow
this app to access it.

The alternative to patching system is decompiling the APK,
changing the targetSdkVersion, recompiling, making it
platform signed, and, in the process, losing Play Store
updates and the ability to use Google Photos as a
PhotosProvider.

Change-Id: Ie8950f209b5289b0825a6178fa00a1ca17a39b38