Revert "Remove UserLifecycleListener deps from SecureSettingsWrapper"
This reverts commit bf2e2dcf4575c254213e9632ff3321b0ec8747a5.

Reason for revert:
While it is highly unlikely, reverting this is the fastest and easiest
way to to demonstrate that this CL is not the real culprit of the SysUI
perf regression discussed in Bug 355334507.

Once it is confirmed that this CL is not the culprit, we will re-land
this as this CL should indeed improve our memory consumption.

Bug: 352354308
Bug: 355334507
Flag: EXEMPT refactor
Test: presubmit
Change-Id: Ib4ce3cb283f5165cff0e66c9a6ddc29910abd911