Avoid harmless NPE in sendResultReceiverFailureLocked()
This is a follow up CL to my previous CL [1], which removed

  ZeroJankProxy.Callback#isInputShownLocked().

While I apparently overlooked scenarios when ResultReceiver is null,
we catch the NPE then only print the stacktrace to logcat [2], which
is why we could not notice it in presubmit.

Let's fix it to avoid confusions. There must be no user-observable
behavior change anyway.

 [1]: Ib719f6018965c3554c8c2d8c2050ad51962d8a25
      f3dbea0b36162a3678cf4e6c1c243b446327a657
 [2]: I008198157424c544d4973eb94434b458e27f15fd
      af4d0786dfad30cd093908365ede58730daaa2d8

Bug: 349904272
Test: presubmit
Flag: android.view.inputmethod.use_zero_jank_proxy
Change-Id: Idf898fb4ca7fdef64e81d92a124f035d244ff3b3