Remove UserLifecycleListener deps from SecureSettingsWrapper (2nd try)
This is the 2nd attempt to submit my original CL [1], which got
reverted [2] to verify whether it was causing jank benchmark
regression discussed in Bug 355334507 or not. As reverting my original
CL did not affect the benchmark, I am now submitting this again.

As part of removing the dependency on UserLifecycleListener from
random utility classes under com.android.server.inputmethod.*, this CL
updates SecureSettingsWrapper so that it does not need to register
UserLifecycleListener by itself.

This is a mechanical refactoring CL. The observable semantics should
remain unchanged.

 [1]: If903fe15e7e71eed2fccb45c082d8efb8176f26c
      bf2e2dcf4575c254213e9632ff3321b0ec8747a5
 [2]: Ib4ce3cb283f5165cff0e66c9a6ddc29910abd911
      b7ca48378c8d6aeb19481d9670078dc6927d38ec

Fix: 352354308
Test: presubmit
Flag: EXEMPT refactor
Change-Id: I7f1502d19df95e6ba5228dd8b501e7e98d210cfd