Flag off parts of SBKVM/KVM#reset.
These calls can cause spurious shade expansion events that cause tests to fail in setup. We are trying to get rid of "reset" as a concept in the refactor, and have views/state that need to be "reset" respond to the actual state changes instead.

Bug: 278086361
Flag: com.android.systemui.keyguard_wm_state_refactor
Test: cts presubmits
Change-Id: I7152f6e49c3396f62f2720195be06d00c0e5ca20