Take task snapshot on exiting PiP
When expanding PiP to full-screen, we first keep Task windowing mode to
be pinned and therefore, the behind Task is going to be invisible and
misses the chance to take the task snapshot.

Fix the issue by forcefully take task snapshot when the Task windowing
mode is toggled.

Flag: EXEMPT bugfix
Video: http://recall/-/aaaaaabFQoRHlzixHdtY/gvNLaUdJmwBpmtJ4VTxqCt
Bug: 342310361
Test: Expand PiP to full-screen with other app in foreground
Change-Id: I8ed6d1b41860c702d71e73d0702e835fcddea6a2