Interface with Repeat keys feature flag and Settings
Add the APIs to check whether repeat keys feature is enabled and to set
the configurable values for the key repeat timeout[1] and key repeat
delay[2] if it is to be set via Settings.
[1] Key repeat timeout is the time duration before which the key starts
repeating on being pressed down.
[2] Key repeat delay is the time interval between successive key repeats
when a key is pressed down.
The minimum and maximum values for both settings are set according the
the corresponding values set in ChromeOS for consistency.

Bug: 363979307
Flag: com.android.hardware.input.keyboard_repeat_keys
Test: Presubmit
Change-Id: I7120cdf54c9c3c65e03751290b94ac767cbef900