Revert "[flexiglass] Adds gestureFilter lambda to STL"
This reverts commit b28dc3dc9dbc567f13870150a2434edebb0f9e77.

Reason for revert: As discussed offline, the system gesture exclusion
was introduced for back gesture, so it should not be used for our
custom use cases. Moreover the gestureFilter API introduced in this CL
probably does not work as we would expect given that we still consume
(but ignore) all pointer events if the gesture is filtered. A better
name for this would have been "consumeButIgnoreGesture: (Offset) ->
Boolean", which highlights why this API should probably not be
introduced in the first place.

Change-Id: I2501b39e0a18b8659578eaac443a12a7a3da5102